refactor(core): extract semantic chunk planning (#1108)
